Extend update protocol to allow adding styles
I want to allow users to inject their own styles into the rendered
markdown. To that end, I took the styles that were already in
`shell.html` and pulled them into `default.css`, and made the update
event a serialized struct that contains the styles and and the content.

Next step will be adding a `--stylesheet` flag to allow including
user-supplied css.