Listen for create events in the file watcher
Vim apparently doesn't write in place to the file, but rapidly deletes
the file and recreates it. As we were only watching for `Write` events,
this meant that files edited in Vim didn't trigger a refresh of the
preview.

This exact situation is discussed in a [Github
issue](https://github.com/passcod/notify/issues/113) in the `notify`
repo. However, contrary to what is mentioned in the comments, we _do_
seem to continue getting events for the file after it has been deleted
and recreated. This means that we can just extend our match pattern
include `Create`s, and that there's no need to watch the parent
directory or use the `PollWatcher` (both solutions discussed in the
issue).